separacao de validacoes em novos modulos js
